Integrate with service to get asset historical prices

+++ fin-life-api/migrations/1684447559123-create_asset_historical_prices_table.ts
@@ -23,7 +23,7 @@ export class createAssetHistoricalPricesTable1684447559123 implements MigrationI
           },
           {
             name: 'closing_price',
-            type: 'float'
+            type: 'int'
           }
         ],
         foreignKeys: [
